Move objectiron to src/dash/parser and make it compliant with factorymaker
